test: move helicone gemini test to tests/litellm/
Move test to correct directory per PR template requirements.
